requirements.txt
streamlit==1.31.0
pandas==2.2.0
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openpyxl==3.1.2
groq==0.4.2
python-dotenv==1.0.1
plotly==5.18.0
requests==2.31.0